Fix: bugs in collisions
The collision had two bugs: after collision with the circle, the bubbles
sometimes floated outside the circle a little, and after collision with
another bubble, the bubbles sometimes got stuck and sometimes moved in
wrong direction.

The solution was proposed by Rabeeb, a batchmate of mine. He told me to
check if the dot product of the normal of the collision surface and the
velocity is positive. If positive, handle collision. Otherwise, ignore.

I don't know how or why this works, but it does seem to work. Oh well.